requirements.txt
python-telegram-bot>=20.0,<21.0
pytz
requests
python-dotenv
nest_asyncio
